requirements.txt
pytorch-ignite>=0.5.0
matplotlib
numpy
pillow
tqdm
tensorboard
scipy   